Add functionality to load an image file
As it turns out loading an SVG file and drawing it to canvas is tricky.

SVG files may or may not have width and height attributes.

If these attributes are missing, drawing to a canvas results in an empty image.

A workaround is to extract the viewbox information and add those as attributes.

Additionally, there are fallback dimensions, if there is also no viewbox.